Feat: Fixed for new metadata
