+++ src/app/proposals/[id]/_components/vote.component.tsx
